Add a CODEOWNERS file
As a way to document who generally should be pulled in to PRs / who people can direct issues or questions to, I've added a CODEOWNERS file that will result in Github showing the relevant person/people at the top of the file. It's generally structured as:

@C0deMunk33: All the actual contracts, tests and migrations
@ashleyvega & @edd: packaging, test running, automations that sit around the interesting parts

Given that the rate of change on this repo is going to be low, consider this more documentation than anything else.